refactor: split native locale resolution helper
